+++ lib/src/mobile_scanner_platform_interface.dart
@@ -146,6 +146,23 @@ abstract class MobileScannerPlatform extends PlatformInterface {
     throw UnimplementedError('getSupportedLenses() has not been implemented.');
   }
 
+  /// Determine the lens type best suited for close-range QR scanning.
+  ///
+  /// Returns the [CameraLensType] whose camera can focus closest to the
+  /// phone. On iOS 15+, this is based on the minimum focus distance reported
+  /// by each camera; on platforms or OS versions where that data is not
+  /// available or not actionable, this returns [CameraLensType.normal].
+  ///
+  /// Returns `null` if the device has no camera for the given [facing]
+  /// direction.
+  Future<CameraLensType?> getBestQrScanningLens({
+    CameraFacing facing = CameraFacing.back,
+  }) {
+    throw UnimplementedError(
+      'getBestQrScanningLens() has not been implemented.',
+    );
+  }
+
   /// Update the scan window to the given [window] rectangle.
   ///
   /// Any barcodes that do not intersect with the given [window] will be